refactor: move validation works tocli.validation
